Fix adding locations to mobile application (#1681)
* Fix for android location adding fails

Added a fix for the file path error when creating locations on Android.

* Check for read-only folders

Well, it's a hacky way, and I'm sure rust may have a better solution, but for now, this checks if the app can make a file called "text.txt". If it works, it's deleted and marked as readable.

* UI Fixes for Navigation

On Android, the back button would not render for some reason. Had to go around and add to the navigation options for each screen the left arrow to go back.

Also, built the about page from the desktop builds of Spacedrive in `about.tsx`.

* Tailwind + Navigation refactor

Navigation has been fixed to not use useEffect, but now exists in the `<Stack.Screen />` code.

Also, refactored `about.tsx` to now use Tailwind instead of pure styles.

* Update AppearanceSettings.tsx

* Update AppearanceSettings.tsx

* Clean up of unused import calls

* Quick Fixes

About now has build info showing up.

Had to remove the read-only check because it was breaking adding new folders.

* Update ImportModal.tsx

Delete unused code.

* Update Settings.tsx

Remove unused imports.